crypto_lib/pyproject.toml
[tool.poetry]
name = "crypto-lib"
version = "0.1.0"
description = ""
authors = ["bsinghvt <baljotsingh1986@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pycryptodomex = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
